TK Mechanic Polish (Push/Pull)
TK Mechanic now has different basic functionality. Press and hold left click to pick up a physics object, and release to throw it. You do not have to hold right mouse to do so, and you cannot move while and object is held. Doors still work with the new system but can be a bit buggy, so some fixes will need to be done for them soon.