Practical Replacement for Boxes/Pause and Main Menu Prototype
Boxes have been given context as batteries, box pressure plates context as sockets. (Using Placeholder Clip Art for now to Denote this) When a battery is connected to the socket, a green light ignites showing the power is on, and the door opens. Placeholder Main Menu and Pause Menus added. Main Menu is in it's own main menu level. (Settings are currently placeholder, as a programmer will most likely need to implement them using console command calls). Pause menu is brought up using the Tab key.